+++ sortition/sortition.go
@@ -5,8 +5,8 @@ import (
 )
 
 func EvaluateSortition(seed VerifiableSeed, signer crypto.Signer, total, threshold int64) (bool, Proof) {
-	index, proof := evaluate(seed, signer, total)
-	if index < threshold {
+	index, proof := evaluate(seed, signer, uint64(total))
+	if int64(index) < threshold {
 		return true, proof
 	}
 
@@ -14,9 +14,9 @@ func EvaluateSortition(seed VerifiableSeed, signer crypto.Signer, total, thresho
 }
 
 func VerifyProof(seed VerifiableSeed, proof Proof, public crypto.PublicKey, total, threshold int64) bool {
-	index, result := verify(seed, public, proof, total)
+	index, result := verify(seed, public, proof, uint64(total))
 	if !result {
 		return false
 	}
-	return index < threshold
+	return int64(index) < threshold
 }
